[#2]refactor: package name group -> classroom
