+++ src/main/java/soma/haeya/lms/classroom/client/ClassroomApiClient.java
@@ -1,10 +1,10 @@
-package soma.haeya.lms.group.client;
+package soma.haeya.lms.classroom.client;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.service.annotation.HttpExchange;
 import org.springframework.web.service.annotation.PostExchange;
-import soma.haeya.lms.group.model.request.ClassroomCreateRequest;
+import soma.haeya.lms.classroom.model.request.ClassroomCreateRequest;
 
 @Component
 @HttpExchange("/v1/classroom")
