+++ src/main/java/soma/haeya/lms/classroom/client/config/ClassroomApiRestClientConfig.java
@@ -1,4 +1,4 @@
-package soma.haeya.lms.group.client.config;
+package soma.haeya.lms.classroom.client.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.client.support.RestClientAdapter;
 import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-import soma.haeya.lms.group.client.ClassroomApiClient;
+import soma.haeya.lms.classroom.client.ClassroomApiClient;
 
 @Configuration
 public class ClassroomApiRestClientConfig {
